+++ Tests/executors/CC/core/src/ChunksExecutor.c
@@ -1,4 +1,4 @@
-#include <APIExecutors/BaseExecutor.h>
+#include <Executors/BaseExecutor.h>
 
 #include <signal.h>
