+++ Tests/executors/CC/core/src/DownloadExecutor.c
@@ -1,4 +1,4 @@
-#include <APIExecutors/BaseExecutor.h>
+#include <Executors/BaseExecutor.h>
 
 #include <Utility/WebFrameworkUtility.h>
