+++ Tests/executors/CC/core/src/HelloExecutor.c
@@ -1,4 +1,4 @@
-#include <APIExecutors/BaseExecutor.h>
+#include <Executors/BaseExecutor.h>
 #include <JSONParser.h>
 
 typedef struct
