+++ Tests/executors/CC/core/src/LocalizationExecutor.c
@@ -1,4 +1,4 @@
-#include <APIExecutors/BaseExecutor.h>
+#include <Executors/BaseExecutor.h>
 
 #include <JSONBuilder.h>
 #include <JSONParser.h>
