+++ Tests/executors/CC/core/src/UploadMultipartExecutor.c
@@ -1,4 +1,4 @@
-#include <APIExecutors/BaseExecutor.h>
+#include <Executors/BaseExecutor.h>
 
 DEFINE_DEFAULT_EXECUTOR(UploadMultipartExecutor, HEAVY_OPERATION_STATELESS_EXECUTOR)
