+++ Tests/executors/CXX/core/include/AssetsExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseStatelessExecutor.hpp>
+#include <Executors/BaseStatelessExecutor.hpp>
 
 class AssetsExecutor : public framework::BaseStatelessExecutor
 {
