+++ Tests/executors/CXX/core/include/ChunksExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseHeavyOperationStatelessExecutor.hpp>
+#include <Executors/BaseHeavyOperationStatelessExecutor.hpp>
 
 class ChunksExecutor : public framework::BaseHeavyOperationStatelessExecutor
 {
