+++ Tests/executors/CXX/core/include/DownloadExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseHeavyOperationStatelessExecutor.hpp>
+#include <Executors/BaseHeavyOperationStatelessExecutor.hpp>
 
 class DownloadExecutor : public framework::BaseHeavyOperationStatelessExecutor
 {
