+++ Tests/executors/CXX/core/include/HelloExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseStatelessExecutor.hpp>
+#include <Executors/BaseStatelessExecutor.hpp>
 
 class HelloExecutor : public framework::BaseStatelessExecutor
 {
