+++ Tests/executors/CXX/core/include/LocalizationExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseStatelessExecutor.hpp>
+#include <Executors/BaseStatelessExecutor.hpp>
 
 class LocalizationExecutor : public framework::BaseStatelessExecutor
 {
