+++ Tests/executors/CXX/core/include/UploadMultipartExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseHeavyOperationStatelessExecutor.hpp>
+#include <Executors/BaseHeavyOperationStatelessExecutor.hpp>
 
 class UploadMultipartExecutor : public framework::BaseHeavyOperationStatelessExecutor
 {
