+++ Tests/executors/CXX/core/include/UploadOctetStreamExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseStatefulExecutor.hpp>
+#include <Executors/BaseStatefulExecutor.hpp>
 
 #include <fstream>
