+++ Tests/executors/CXX/database/include/CRUDExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseStatelessExecutor.hpp>
+#include <Executors/BaseStatelessExecutor.hpp>
 
 class CRUDExecutor : public framework::BaseStatelessExecutor
 {
