+++ Tests/executors/CXX/database/include/MultiUserExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseHeavyOperationStatefulExecutor.hpp>
+#include <Executors/BaseHeavyOperationStatefulExecutor.hpp>
 
 class MultiUserExecutor : public framework::BaseHeavyOperationStatefulExecutor
 {
