+++ Tests/executors/CXX/load_balancer/include/IdExecutor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <APIExecutors/BaseHeavyOperationStatelessExecutor.hpp>
+#include <Executors/BaseHeavyOperationStatelessExecutor.hpp>
 
 class IdExecutor : public framework::BaseHeavyOperationStatelessExecutor
 {
